+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Fixed `Folder::move()` deleting the folder if you move folder into itself, created empty file instead
     * Fixed moving `Flex Page` to itself causing the page to be lost [#3227](https://github.com/getgrav/grav/issues/3227)
     * Fixed `PageStorage` from detecting files as pages
+    * Fixed `UserIndex` not implementing `UserCollectionInterface`
 
 # v1.7.9
 ## 03/19/2021